Deleting unnecessarily components
