+++ fe-library-app/src/components/AppRouter/AppRouter.tsx
@@ -2,7 +2,6 @@ import { Navigate, Route, Routes } from 'react-router'
 
 import HomePage from './pages/HomePage'
 import LoginPage from './pages/LoginPage'
-import ProfilePage from './pages/ProfilePage'
 import PrivateRouter from './PrivateRouter'
 import CreateUpdatePage from './pages/CreateUpdatePage'
 import { isUserLoggedIn } from '../../services/AuthService'
@@ -14,7 +13,6 @@ const AppRouter = () => {
       <Route path='/' element={<Navigate to="/home" />} />
       <Route path='/home' element={<HomePage />} />
       <Route element={<PrivateRouter/>}>
-        <Route path='/profile' element={<ProfilePage />}/>
         <Route path='/add_modify' >
           <Route index element={<CreateUpdatePage />} />
           <Route path=':id' element={<CreateUpdatePage />} />
