+++ fe-library-app/src/components/CreateUpdateForm/CreateUpdateBookForm.module.css
@@ -2,6 +2,8 @@
 .add_modify_book_form{
     display: flex;
     flex-direction: column;
+    height: 90%;
+    overflow-y: scroll;
 }
 
 .add_modify_book_form input,
@@ -76,6 +78,10 @@
 
 .add_author_form{
     display: none;
+}
+
+.add_modify_book_button{
+    margin-bottom: 10px;
 }
 
 @media only screen and (min-width: 768px) {
