--- fe-library-app/src/components/SideBar/SideBar.module.css
@@ -1,51 +1,0 @@
-.sideBar {
-    display: flex;
-    bottom: 0;
-    right: 0;
-    position: fixed;
-    min-height: 100vh;
-    width: 50%;
-    z-index: 90;
-    background-color: var(--dark_color);
-    color:white;
-    opacity: 0.9;
-    flex-direction: column;
-    justify-content: flex-end;
-    padding-bottom: 100px;
-}
-
-.sideBar button{
-    display: none;
-}
-
-.sideBar a{
-    color: white
-}
-
-.sideBarHidden {
-    display: none;
-}
-
-@media only screen and (min-width: 768px){
-    .sideBar{
-    display: block;
-    top: 170px;
-    bottom: 0;
-    left: 0;
-    position: fixed;
-    min-height: 100vh;
-    width: var(--sidebar_width);
-    background-color: var(--dark_color);
-    color:white;
-    }
-
-    .sideBar a{
-        display: none;
-    }
-
-    .sideBar button{
-        all: unset;
-        text-decoration: underline;
-        cursor: pointer;
-    }
-}
